refactor: 유효기간 훅 validation 로직 분리​​
